requirements.txt
Django>=2.2
pyyaml
requests
django-embed-video